Thumbs up icon
Removed a a text counter and left it as a number counter. I then stacked the number of likes element on top of the thumbs up Icon. I had to adjust the styling of the likes counter to fit the user story and change the position of the number on the thumbs up icon.